настроил линтеры и форматирование и часть верстки
