--- index.html
@@ -1,55 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>card-game</title>
-    <link rel="stylesheet" href="style.css" />
-  </head>
-  <body>
-    <div id="app"></div>
-  </body>
-  <script src="script.js"></script>
-</html>
-
-
-
-
-
-<!-- <div class="body game">
-    <div class="game__header">
-        <div class="time">
-            <div class="time__block">
-                <p class="time__name">min</p> 
-                <p class="min__value">00</p>
-            </div>
-            <div class="time__block">
-                <p class="time__name">sek</p> 
-                <p class="sek__value">.00</p>
-            </div>
-        </div>
-        <button class="button-start">Начать заново</button>
-    </div>
-    <ul class="game__zone">
-        <li class="game__card">
-            <img class="game-card" src="img/закрытая.png" alt="1">
-        </li>
-        <li class="game__card">
-            <img class="game-card" src="img/закрытая.png" alt="2">
-        </li>
-        <li class="game__card">
-            <img class="game-card" src="img/закрытая.png" alt="3">
-        </li>
-        <li class="game__card">
-            <img class="game-card" src="img/закрытая.png" alt="4">
-        </li>
-        <li class="game__card">
-            <img class="game-card" src="img/закрытая.png" alt="5">
-        </li>
-        <li class="game__card">
-            <img class="game-card" src="img/закрытая.png" alt="6">
-        </li>
-    </ul>
-</div> -->
-
