+++ src/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8" />
+        <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>card-game</title>
+        <link rel="stylesheet" href="style.css" />
+    </head>
+    <body>
+        <div id="app"></div>
+    </body>
+    <script type="module" src="./script.js"></script>
+</html>
