+++ src/style.css
@@ -1,0 +1,170 @@
+@font-face {
+    font-family: "StratosSkyeng";
+    font-style: normal;
+    font-weight: 400;
+    font-display: swap;
+    src: url("./fonts/stratosskyengweb-regular.woff2") format("woff2");
+}
+
+* {
+    margin: 0;
+    padding: 0;
+}
+
+.body {
+    background-color: #004980;
+    height: 100vh;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.start {
+    width: 480px;
+    height: 459px;
+    background: #c2f5ff;
+    border-radius: 12px;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    align-items: center;
+}
+
+.start__title {
+    font-family: "StratosSkyeng";
+    font-style: normal;
+    font-weight: 400;
+    font-size: 40px;
+    line-height: 48px;
+    text-align: center;
+    font-feature-settings: "pnum" on, "lnum" on;
+    color: #004980;
+}
+
+.levels {
+    display: flex;
+    gap: 24px;
+    margin-bottom: 16px;
+}
+
+.form_radio_btn {
+    display: inline-block;
+}
+
+.level-input {
+    display: none;
+}
+
+.level-label {
+    display: inline-block;
+    cursor: pointer;
+    width: 98px;
+    height: 98px;
+    font-family: "StratosSkyeng";
+    background: #ffffff;
+    border-radius: 12px;
+    font-style: normal;
+    font-weight: 400;
+    font-size: 64px;
+    line-height: 72px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    color: #0080c1;
+    border: none;
+}
+
+/* Checked */
+.level-input:checked + label {
+    background: #ffe0a6;
+}
+
+/* Hover */
+.level-label:hover {
+    opacity: 0.7;
+}
+
+.button-start {
+    width: 246px;
+    height: 50px;
+    background: #7ac100;
+    border-radius: 12px;
+    font-family: "StratosSkyeng";
+    font-style: normal;
+    font-weight: 400;
+    font-size: 24px;
+    line-height: 32px;
+    font-feature-settings: "pnum" on, "lnum" on;
+    color: #ffffff;
+    border: none;
+}
+
+.button-start:hover {
+    opacity: 0.8;
+}
+
+.button-start:active {
+    opacity: 1;
+}
+
+.game {
+    display: flex;
+    flex-direction: column;
+}
+
+.game__header {
+    width: 90vw;
+    max-width: 979px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    flex-wrap: wrap;
+    margin-bottom: 46px;
+}
+
+.time {
+    font-family: "StratosSkyeng";
+    font-style: normal;
+    font-weight: 400;
+    font-size: 64px;
+    line-height: 72px;
+    color: #ffffff;
+    display: flex;
+}
+
+.time__block {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-end;
+}
+
+.time__name {
+    font-family: "StratosSkyeng";
+    font-style: normal;
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 32px;
+    font-feature-settings: "pnum" on, "lnum" on;
+    color: #ffffff;
+}
+
+.game__zone-1 {
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr;
+    gap: 15px;
+    list-style: none;
+}
+
+.game__zone-2 {
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr 1fr;
+    gap: 15px;
+    list-style: none;
+}
+
+.game__zone-3 {
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr 1fr 1fr 1fr;
+    gap: 15px;
+    list-style: none;
+}
